perf: change game id to long and drop igdb id

+++ Interfaces/IGameService.cs
@@ -2,6 +2,6 @@ using Hatsu.Models;
 
 namespace Hatsu.Interfaces;
 
-public interface IGameService : IService<Game, int>
+public interface IGameService : IService<Game, long>
 {
 }
