+++ Models/Entry.cs
@@ -7,7 +7,7 @@ public class Entry : Entity<int>
     public int UserId { get; set; }
     public User User { get; set; } = null!;
 
-    public int GameId { get; set; }
+    public long GameId { get; set; }
     public Game Game { get; set; } = null!;
 
     public EntryStatus Status { get; set; } = EntryStatus.Backlog;
