+++ Services/GameService.cs
@@ -3,7 +3,7 @@ using Hatsu.Models;
 
 namespace Hatsu.Services;
 
-public class GameService : Service<Game, int>, IGameService
+public class GameService : Service<Game, long>, IGameService
 {
     public GameService(IGameRepository pRepository) : base(pRepository) { }
 }
